+++ examples/03-route-guards/src/app/modules/auth/auth.guard.ts
@@ -2,10 +2,11 @@ import { CanActivate, injectable, RequestContext } from '@ditsmod/core';
 
 @injectable()
 export class AuthGuard implements CanActivate {
+  constructor(private ctx: RequestContext) {}
   /**
    * Here you need implement more logic.
    */
-  async canActivate(ctx: RequestContext, params: any[]) {
+  async canActivate(params: any[]) {
     return false;
   }
 }