+++ examples/10-openapi/src/app/modules/routed/second/bearer.guard.ts
@@ -1,4 +1,4 @@
-import { CanActivate, Req, RequestContext, Status } from '@ditsmod/core';
+import { CanActivate, RequestContext, Status } from '@ditsmod/core';
 import { JwtService, VerifyErrors } from '@ditsmod/jwt';
 import { oasGuard } from '@ditsmod/openapi';
 
@@ -20,10 +20,10 @@ import { oasGuard } from '@ditsmod/openapi';
   },
 })
 export class BearerGuard implements CanActivate {
-  constructor(private jwtService: JwtService) {}
+  constructor(private jwtService: JwtService, private ctx: RequestContext) {}
 
-  async canActivate(ctx: RequestContext) {
-    const authValue = ctx.nodeReq.headers.authorization?.split(' ');
+  async canActivate() {
+    const authValue = this.ctx.nodeReq.headers.authorization?.split(' ');
     if (authValue?.[0] != 'Token') {
       return false;
     }
@@ -35,7 +35,7 @@ export class BearerGuard implements CanActivate {
       .catch((err: VerifyErrors) => false as const); // Here `as const` to narrow down returned type.
 
     if (payload) {
-      ctx.req.jwtPayload = payload;
+      this.ctx.req.jwtPayload = payload;
       return true;
     } else {
       return false;