+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
+<a name="core-2.33.1"></a>
+# [core-2.33.1](https://github.com/ditsmod/ditsmod/releases/tag/core-2.33.1) (2023-01-07)
+
+### Bug fixes
+
+- Fixed `CanActivate` interface.
+
 <a name="core-2.33.0"></a>
-# [core-2.33.0](https://github.com/ditsmod/ditsmod/releases/tag/core-2.33.0) (2023-01-07)
+## [core-2.33.0](https://github.com/ditsmod/ditsmod/releases/tag/core-2.33.0) (2023-01-07)
 
 ### BREAKING CHANGES